Fixed the toe bone flipping out of place
In the process I moved the connection point for the toe up as Tob suggested he
would normally do.  I probably still need to adjust the how the foot rotates as
the front of the foot bone drops down a little when rolling forward.  But this
is much better than the previous problem of the toe bone flipping around the
Y-axis to be upside down under the foot in certain positions.